requirements.txt
playwright==1.58.0
python-dotenv==1.2.1
Flask==3.0.0
Flask-Login==0.6.3
bcrypt==4.3.0
graphifyy
